third_party/WebKit/LayoutTests/editing/deleting/delete-3800834-fix.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<div id="root" contenteditable='true'>
<span id="test">Foo<BR><BLOCKQUOTE>Bar</BLOCKQUOTE></span>
</div>
<div id="log"></div>
<script>
test(function() {
    var selection = window.getSelection();
    selection.collapse(document.getElementById('test').firstChild, 0);
    for (var i = 0; i < 8; i++)
        selection.modify("move", "forward", "character");
    for (var i = 0; i < 4; i++)
        document.execCommand('delete');
    assert_equals(document.getElementById('root').innerHTML, '\n<span id="test">Foo</span>');
});
</script>
